requirements.txt
appdirs==1.4.4
asgiref==3.2.10
backcall==0.2.0
bcrypt==3.1.7
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9.2
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20181111
distlib==0.3.1
Django==3.0.7
django-appconf==1.0.4
django-cors-headers==3.4.0
django-crispy-forms==1.9.1
django-crontab==0.7.1
django-filter==2.3.0
django-formtools==2.2
django-import-export==2.2.0
django-reversion==3.0.7
django-simpleui==2020.7
django-staticfiles==1.2.1
djangorestframework==3.11.0
drf-yasg==1.17.1
et-xmlfile==1.0.1
filelock==3.0.12
future==0.18.2
httplib2==0.9.2
idna==2.9
importlib-metadata==2.0.0
inflection==0.5.0
ipython==7.15.0
ipython-genutils==0.2.0
itypes==1.2.0
jdcal==1.4.1
jedi==0.17.0
Jinja2==2.11.2
MarkupPy==1.14
MarkupSafe==1.1.1
mysqlclient==1.4.6
odfpy==1.4.1
openpyxl==3.0.3
packaging==20.4
paramiko==2.7.1
parso==0.7.0
pbr==5.5.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.6.1
PyMySQL==0.9.3
PyNaCl==1.4.0
pyparsing==2.4.7
PySocks==1.7.1
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.15.0
sqlparse==0.3.1
stevedore==3.2.2
tablib==2.0.0
traitlets==4.3.3
uritemplate==3.0.1
urllib3==1.25.9
virtualenv==20.1.0
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
wcwidth==0.2.4
WebOb==1.8.6
xadmin==2.0.1
xlrd==1.2.0
xlwt==1.3.0
zipp==3.4.0
